coreaudio: Add function to get number of devices
We can get the number of devices by getting the size of the
kAudioHardwarePropertyDevices property in bytes and dividing
it by the size of an `AudioDeviceId` (uint32).

note: I couldn't figure out how to import the
audio-object-property-id type into `coreaudio.lisp`. It would
be ideal to reuse the declaration in `coreaudio-cffi.lisp`, but
I'm not sure if this is possible with ctypes.